chore : setting수정
