+++ jsconfig.json
@@ -6,15 +6,17 @@
     "paths": {
       "@/*": ["src/*"] // '@' 별칭을 'src' 폴더로 설정 (모듈 경로 참조 시 간단하게 사용 가능)
     },
-    "types": ["vite/client"] // Vite 프로젝트에서 자동 완성 및 타입 정보를 제공 (Vite 관련 타입을 포함)
+    "types": ["vite/client"], // Vite 프로젝트에서 자동 완성 및 타입 정보를 제공 (Vite 관련 타입을 포함)
+    "ignoreDeprecations": "6.0"
   },
   "include": [
     "src/**/*.d.ts", // 'src' 폴더 내에 있는 모든 타입 정의 파일 포함
     "src/**/*.js", // 'src' 폴더 내에 있는 모든 JavaScript 파일 포함
     "src/**/*.jsx", // 'src' 폴더 내에 있는 모든 JSX 파일 포함 (React 파일)
     "src/**/*.ts", // 'src' 폴더 내에 있는 모든 TypeScript 파일 포함
-    "src/**/*.tsx" // 'src' 폴더 내에 있는 모든 TSX 파일 포함 (React와 TypeScript를 같이 사용할 때)
-, "src/skeletons/SkeletonMovieDetail.jsx"  ],
+    "src/**/*.tsx", // 'src' 폴더 내에 있는 모든 TSX 파일 포함 (React와 TypeScript를 같이 사용할 때)
+    "src/skeletons/SkeletonMovieDetail.jsx"
+  ],
   "exclude": [
     "node_modules", // 'node_modules' 폴더는 검사하지 않음
     "**/node_modules", // 서브 디렉토리의 'node_modules'도 검사하지 않음
