+++ src/components/Loading.jsx
@@ -1,0 +1,7 @@
+export default function Loading() {
+  return (
+    <div className="absolute top-0 left-0 flex h-full w-full flex-col items-center justify-center">
+      <div className="text-center text-3xl">🌀Loading</div>
+    </div>
+  );
+}
